+++ autopush/websocket.py
@@ -405,18 +405,14 @@ class SimplePushServerProtocol(WebSocketServerProtocol):
         self.ps.metrics.timing("client.socket.lifespan", duration=elapsed,
                                tags=self.base_tags)
 
-        # Log out the connection close and info
-        log.msg("Connection Close", code=code, reason=reason,
-                was_clean=wasClean, lifespan=elapsed,
-                ping_time_out=self.ps.ping_time_out)
-
         # Cleanup our client entry
         if self.ps.uaid and self.ap_settings.clients.get(self.ps.uaid) == self:
             del self.ap_settings.clients[self.ps.uaid]
 
-        # Cancel any outstanding deferreds
+        # Cancel any outstanding deferreds that weren't already called
         for d in self.ps._callbacks:
-            d.cancel()
+            if not d.called:
+                d.cancel()
 
         # Attempt to deliver any notifications not originating from storage
         if self.ps.direct_updates: